back-end/requirements.txt
alembic==1.3.2
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
httpie==1.0.3
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
yapf==0.29.0
